Divide operation + float operand + test divide and float

+++ src/main/java/rpn/CLI.java
@@ -12,7 +12,7 @@ public class CLI {
         String expression = Stream.of(args).collect(Collectors.joining(" "));
 
         System.out.println("About to evaluate '" + expression + "'");
-        long result = 0;
+        double result = 0;
         try {
             result = evaluate(expression);
         } catch (UnsupportedExpressionException | UnsufficientArgumentException e) {
@@ -21,7 +21,7 @@ public class CLI {
         System.out.println("> " + result);
     }
 
-    static long evaluate(String expression) throws UnsupportedExpressionException, UnsufficientArgumentException {
+    static double evaluate(String expression) throws UnsupportedExpressionException, UnsufficientArgumentException {
         return Interpreter.evaluate(expression);
     }
 }
